Call changeSessionData action only if the headers are changed
