+++ src/actions/fetchTodayData.js
@@ -1,20 +1,6 @@
 import axios from 'axios';
-import { CHANGE_SESSION_DATA, FETCH_TODAY_DATA } from '../actionTypes';
-import saveToSessionStorage from '../utils/sessionStorage';
-
-const changeSessionData = data => {
-  saveToSessionStorage(data, true);
-  return {
-    type: CHANGE_SESSION_DATA,
-    payload: {
-      uid: data.uid,
-      accessToken: data['access-token'],
-      tokenType: data['token-type'],
-      expiry: data.expiry,
-      client: data.client,
-    },
-  };
-};
+import { FETCH_TODAY_DATA } from '../actionTypes';
+import changeSessionData from './changeSessionData';
 
 const pureFetchTodayData = data => ({
   type: FETCH_TODAY_DATA,
@@ -34,7 +20,9 @@ const fetchTodayData = ({
     expiry,
   },
 }).then(response => {
-  changeSessionData(response.headers);
+  if (response.headers['access-token'] !== '') {
+    changeSessionData(response.headers);
+  }
   dispatch(pureFetchTodayData(response.data));
 });
